refactor: move executeSeedFile to vite-plugin-db
